Start work on events.
I added some logic that lets you do (e.g.):

    <button clicked="go" />

where `go` is a method on the data context. Right now the method has to
have the same prototype as the rest of the efl's callbacks, and there's
a discussion to be had as to what the right thing to do there is -
passing the efl object breaks some abstraction boundaries.

The example in `_test.py` doesn't work, unfortunately, since the context
of the button with the clicked callback is first, not the root of the
data. We need to pin down how situations like that are going to be
handled; clearly this design needs some more thought.