Start separating widgets and frames.
The code is in a broken state right now; some parts have to be updated
to make it work again.

This is the beginnings of an effort to separate widgets from their
frames. Frames are a structural specification, which does not change
based on run-time events. Widgets, on the other hand, correspond to
actual graphical components. They are generated from the frame/spec,
which should not be polluted by state of a particular instance.